Added HAMFIs for 2010-2014 CHAS
Chas_IncomeLimits_17 is the primary HAMFI file. For each MSA or county equivalent, the file includes the 50% (very low income) limit and 80% (low income) limit. Both figures are based on a 4 person household. 

Census matches these files by geography to each household in the ACS microdata. HAMFIs are adjusted by household size, adding 8% for each person over 4 and subtracting 10% for each person under 4. Additional income thresholds are computed based on the 50% and 80% limits. All thresholds above 80% are computed based on the 80% threshold. All thresholds below 80% are computed based on the 50% threshold. For example, the 30% threshold is computed as .6 times the 50% threshold. The 120% threshold is computed as 1.5 times the 80% threshold.